devel/gvfs: Depend on sysutils/lsof at run time
When a drive cannot be unmounted and returns EBUSY, gvfs calls
"lsof -t /mountpoint" to find which processes have files open.
This list is sent over the "show-processes" signal, which allows
file managers to show which apps are preventing the unmount.

For this to work, sysutils/lsof needs to be around.

PR:		254322
Approved by:	desktop (fluffy)
Differential Revision:	https://reviews.freebsd.org/D32018