security/tcpcrypt: use new divert(4) API
The upstream has been idle for 8 years, so unlikely they will accept
a patch in a near future. John-Mark (maintainer) says the project is
abandoned and he plans to remove the port. Before this happens let
the last version of the port work with modern API, at least to
facilitate an exp-run.  Pull the patch from my github fork.

Approved by:	mikael