devel/freebsd-gcc12: Ignore on FreeBSD 12.
The powerpc flavor fails to build on FreeBSD 12.x as GCC does not
install the secureplt.h header for that version (GCC assumes secureplt
is only used on 13.0 and later).  In addition, the fixes needed to
build FreeBSD with GCC 12.x are not going to be backported to
stable/12, so just skip this port on FreeBSD 12.