devel/spark: fix multiple issues
The devel/spark port had been resurrected by Neel Chauhan based on
PR 266484, but that PR had become stale and did not follow current
rules and conventions (and had some issues that needed to be fixed).

Neel has run out of time fixing the issues and I have taken over and
finished the patches he had been working on.

This updated port has been tested with poudriere testport, but I do
not have a suitable test environment to run functional tests on.

Please address run-time issues that are detected to both the
maintainer of the port and to me.

Approved by:	portmgr (blanket)