Mk/Features/lto.mk: Add checks for powerpc64 and riscv64
Add checks for powerpc64 and riscv64 since they won't work with LTO.

This is a bit crude on powerpc64 but can be improved upon later.
LTO does work on this platform using GCC but will fail using lld.

This commit also introduces LTO_DISABLE_CHECK which can be used to
override.

Reviewed by:	pkubaj (private email)