www/dolibarr15: New port: Modern software package to manage your company or foundation
Dolibarr ERP & CRM is a modern software package to manage your organization's
activity (contacts, suppliers, invoices, orders, stocks, agenda)

It's an Open Source Software (written in PHP language) designed for small,
medium or large companies, foundations and freelances.

You can freely use, study, modify or distribute it according to its Free
Software licence.

You can use it as a standalone application or as a web application to be able
to access it from the Internet or a LAN. (15.x version)