devel/kconfig-frontends: Kconfig files parser and frontend
Kconfig frontends and parser. Port created for NuttX RTOS development.
Kconfig is the configuration language used by the Linux kernel. This package
is a simple copy of the frontends and the parser found in the Linux kernel
source tree, with very minor changes to adapt them to being built out of
the kernel build infrastructure. Kconfig is also used in embedded development.
This program has currently no upstream maintainer. NuttX copy is used.
There are two forks: Debian based on 4.11, Espressif based on version 4.6.

PR:		259256