archivers/py-borgbackup: clean environment for self-tests
This avoids build failures when, for instance, BORG_NEW_PASSWORD
is set to another password than the one used by the self-tests.

Reported by:	SimpleRezo <simplerezo@gmail.com>
PR:		266821

Not bumping PORTREVISION because this only affects the execution
of self-tests, which is pass (then the change is irrelevant) or fail
(then there is no package or installed port and PORTREVISION would
serve no purpose).