security/acme.sh: Move cron example to EXAMPLESDIR
Instead of installing the sample crontab to etc/cron.d, install it to
EXAMPLESDIR

etc/cron.d/acme.sh will still be installed (if not present) but now
contains no active lines and more complete instructions.

Both the cron the newsyslog examples are only installed if the EXAMPLES
config option is ON (default).

Reported by:	bapt