security/john: try to fix the port's build on AArch64
In commit 1e6670382fbb, the observed failure on PowerPC had been
mended in a way that broke AArch64 build.  Robert's investigation
had shown that <altivec.h> should've been included elsewhere, and
it actually is, but guarded by `#ifdef __linux__' for some reason.

Alexander explained that at the time the AltiVec support was just
for two platforms: macOS (first) and Linux (added later).  If the
guard of `#include <altivec.h>' was needed, then we could probably
replace the `#ifdef __linux__' with `#ifndef __APPLE__' for the
same effect on old macOS versions which do not need the #include.

PR:		266732
Submitted by:	Robert Clausecker
Discussed with:	Solar Designer