net/openldap2[4-6]-server: suppress warning message when chown failed.
On UFS with sujournal enabled, the journal is immutable and chown would
fail. This would cause the script to issue an error when a separate UFS
file system is used as OpenLDAP data store, which is not actionable and
causes confusion.

Reported by:	Olivier Nicole <on cs.ait.ac.th>
PR:		ports/266811