audio/mousai-devel: add new port
Discover songs you are aching to know with an easy-to-use interface.

Mousai is a simple application that can identify songs similar to Shazam. Just
click the listen button, and then wait a few seconds. It will magically return
the title and artist of that song!

Note: This uses the API of audd.io, so it is necessary to log in to their site
to get more trials.

Why you will love Mousai?
* Identify the title and artist of the song within seconds.
* Use your microphone or audio from desktop.
* Store the identified song including the album art in history.
* Preview the identified song with the native player.
* Browse and listen the song from different providers.
* Easy-to-use user interface.
* User-friendly keyboard shortcuts.

https://github.com/SeaDve/Mousai