editors/ghostwriter: Upstream project moved under KDE umbrella
- Update home page
- Update distfile fetching

Old URLs are now invalid, although a redirection is being kept
active for the home page, but not for the distfile, which was being
fetched by the FreeBSD cache.

The distfile contents are identical to the old ones, and the produced
package is exactly the same, so no need for a PORTREVISION bump.