www/elinks: fix detection of execinfo
When moving from autotools to autotools to meson, the check from
the execinfo header was lost, readd this test.

When execinfo is not detected, the replacement code instead of error
is an abusive end less loop, while here patch out this part [1].

Note that upstream merge the removal of the code.

While here remove a useless macro which wasn't functional anyway due to
a typo

PR:	267794 [1]
Submitted by:	Dustin Marquess <jailbird@fdf.net> [1]