net-mgmt/smokeping: switch from GITHUB to MASTER_SITES
During the release process, the Smokeping team updates
the version number displayed in the webUI in the archive
files but never report this update on the git repo.
Instead of trying to fix this issue in the port, we move
back to static archives.

Apart from version no difference was found on the code.

Since I'm here, improve the port itself to make linters
happy.

PR:		267815
Reported by:	Gasol Wu <gasol.wu@gmail.com>