games/bass: Add desktop entry; cleanup
Update MASTER_SITES; distfile was un-fetchable.

Give LICENSE a better name and description. Extract and format only
the relevant part from readme.txt and install that as LICENSE_FILE.

Remove pkg-message.in which just reiterated the instructions already
stated in readme.txt. A desktop entry has been added to facilitate
launching the game and to support the locales for which the game has
subtitles.