games/fotaq: Add translations (de, fr, he, it)
Add French, German, Hebrew (subtitles), and Italian (subtitles)
language support as options. Unfortunately, these translations are
all contained in separate distfiles and cannot be simultaneously
installed without creating separate ports or flavors. Since this is an
old game that one likely plays through once, separate ports are probably
not warranted. The original English verison is still the default. It is
noted that the translated version nubmers are different than the English
version, but they are highly unlikely to receive future updates.

Install the entire license text instead of just a stub.

Remove pkg-message in favor of a desktop entry to easily launch the
game from an XDG-compliant environment.

Remove the 2-line pkg-plist in favor of PLIST_FILES and PORTDOCS to
facilitate the transition to multiple distfiles with slightly differing
file structures.

Pet portlint(1).