lang/ruby30: Fix plist error when both CAPIDOCS and DOCS options are on.
When the both options are on, post-install-DOCS-on target is called
after post-install-CAPIDOCS-on has completed. And in the former files
under ${WRKSRC}/doc/ are copied to ${STAGEDIR}${RUBY_DOCDIR}. This
means files under ${WRKSRC}/doc/capi are copied to
${STAGEDIR}${RUBY_DOCDIR}/capi again. And of course
${WRKSRC}/doc/capi/.timestamp is included in them. As a result, even
though ${STAGEDIR}${RUBY_DOCDIR}/capi/.timestamp is removed in
post-install-CAPIDOCS-on target, it is detected as orphaned file at
check-plist phase. So fix it by adding 3rd argument to
${COPYTREE_SHARE} command so .timestamp is excluded from files to be
copied.

PR:		267775
Approved by:	meta (ruby@)
Fixes:		93e3548c57a4 lang/ruby30: Simplify CAPIDOCS handling