x11-toolkits/soqt: switch master site
The Bitbucket distribution point is no more, and the main repository
now lives on Github. At the same time, the auto-generated tarballs
over there cannot be used (they lack the submodules required for
building, and upstream documents that over here:
  https://github.com/coin3d/soqt/releases
Just point MASTER_SITES at the correct tarball. On the upside, we
now get the gzipped file (still, the tarball is a little bit larger
as it now contains some mercurial leftovers, but it is the same for
all relevant content - I checked).
While here, add glu to USE_GL as advised by testport.