devel/lightning: Fix build on armv6/armv7
Fix the problem that made the build fail with LLVM.  However,
compiling with LLVM still leads to test suite failures on ARM
so we compile with gcc instead.  The patch for LLVM is left
in for if a future version of the library fixes the problem.

PR:		268079
Reported by:	fuz@fuz.su
MFH:		2022Q4

(cherry picked from commit fc86afb83def6fe03212dea1ad2e953daa735847)