lang/julia: unbrake after update of math/suitesparse-cholmod
Looks like julia is incompatible with the latest version of suitesparse-cholmod.
suitesparse is an optional dependency so remove it for now and will
add it back when a version of julia that is compatible with it is
released.

PR:		268025
Reported by:	rizor.reardon@yahoo.com (maintainer)

(cherry picked from commit 8a3c1b4542340f4ac480a55ce225b10f10238436)