www/firefox: update to 108.0 (rc1)
Release Notes (soon):
  https://www.mozilla.org/en-US/firefox/108.0/releasenotes/

Drops cubeb patch - we get profiling (of anyone wants to try that),
the code has changed enoughj that it became a hassle. Maybe the
legacy backends have improved, or everyone has finally moved off that.

Harfbuzz dependency is declared at a lower version than it should
be - this is for quarterly's sake, which is already quite stale.
FIXME: next year