x11-toolkits/qt5-gui: minimal plugin independent of option X11
The "minimal" plugin is always needed (or just **one** of
the plugins, but this will do) to allow QtGui applications
to function. Turning off X11 turned them all off (I supposed
"if you're not using X11, you must install the Wayland bits
instead" would work, but that's hard to express in ports).

So remove the plist-sub from the minimal plugin -- there is
enough X11 pulled in regardless of the X11 option, that
the plugin builds (and links against X11 libraries, so
the option is a lie).

The linked PR is not for this port, but demonstrates a problem:
if X11 is turned off, there is no platform plugins at all,
and then there is no way to run Qt-Gui applications (even with
minimal / no graphics). For "command-line" Qt-Gui applications
like qhelpgenerator, that is a problem.

PR:		267949