net-im/mastodon: Resurrect
Mastodon is an alternative implementation of GNU social, a free,
open-source social network.  It is a decentralized alternative to
commercial platforms based on ActivityStreams, Webfinger, PubsubHubbub
and Salmon.  The project focuses on creating a clean REST API and a good
user interface.  Ruby on Rails is used for the back-end, while React.js
and Redux are used for the dynamic front-end.

For easier maintenance of the port some part of the installation
procedure remains to be completed by the end user. The main howto is
available at https://wiki.freebsd.org/Ports/net-im/mastodon.

A great thanks to Stefano Marinelli for the fabulous post on running
Mastodon on a FreeBSD instance which helped a lot in porting this. The
link to the original post is also available on the above mentioned wiki
page.

Sponsored by:	Nepustil