devel/pecl-eio: Update version 2.0.4=>3.00RC4
This was marked for deletion due to EOL of php 7.4 However there was an
updated development version which supports latest php releases. As the
maintainer did not submit any updates hence the port is updated without
actual mantainer's approval but with portmgr blanket approvals.

This was done in consideration to the fact that it's easier to update
rather than rip and resurrect.

Approved by:	portmgr (blanket infrastructure)
Relnotes:	https://pecl.php.net/package-changelog.php?package=eio
Sponsored by:	Bounce Experts