KDE: Update KDE Frameworks to 5.101
Saturday, 10 December 2022

KDE today announces the release of KDE Frameworks 5.101.0.

KDE Frameworks are 83 addon libraries to Qt which provide a wide variety
of commonly needed functionality in mature, peer reviewed and well
tested libraries with friendly licensing terms. For an introduction see
the KDE Frameworks release announcement.

This release is part of a series of planned monthly releases making
improvements available to developers in a quick and predictable manner.

PR:		268291
Exp-run by:	antoine