textproc/py-wcmatch: update to 8.4.1
The upstream package nolonger uses python setuptools, instead building
a wheel using hatchling.  There is no build target to install the
python module directly -- the generated wheel has to be processed by
pip.  Indeed, with the PEP-517 support in recent versions of pip, the
whole package can be built and installed in one commandline using pip.

There are two approaches here:

  * maintain our own setup.py etc. to recreate the setuptools support
    from older versions -- an approach taken in
    textproc/py-pymdown-extensions, which is another package by the
    same upstream author with similar tooling.

  * alternatively, embrace the new python build system, using
    hatchling to create a python wheel and pip to install it to
    staging.

I chose the latter, but it is trickier than it seems.  pip is not
really designed for installing to a staging location as required when
building a port.  pip options for manipulating the installation prefix
(--root, --prefix) pollute the installed _pycache_ files with the
staging directory path.  In fact, this only works at all, because
everything in this package is installed under PYTHON_SITELIBDIR and we
can set the `--target` option to ${STAGING}${PYTHON_SITELIBDIR} to
achive the desired effect.

The PEP-517 ideal would be to use pip to build and install everything,
and manage all the dependency resolution etc., but that's not
compatible with port building.  Fortunately, it is possible to turn
off all of the unrequired functionality.

It would be cleaner to implement a python module to directly convert
wheels into FreeBSD pkgs, or vice-versa.

Changes:	https://github.com/facelessuser/wcmatch/releases/tag/8.4
Changes:	https://github.com/facelessuser/wcmatch/releases/tag/8.4.1