graphics/colord: Update to 1.4.6
This update adds many options including COMPLETIONS, DOCS, MANPAGES,
SANE, and an argyllcms SENSOR option.

Significantly, it removes the docbook-utils dependency, which brought
the entire LaTeX system with it. This unfortunate dependency added
literally hours to the build time of colord and everything which
required it (in particular, gtk30 and the entire GNOME DE).

The comical part is that docbook-utils haven't actually been used by
colord in nearly 5 years; our port was simply never updated to remove
that dependency. Mercifully, this is now fixed.

Many thanks to Antoine for running multiple exp-runs caused by me not
having my act together.

PR:		267239