www/drupal10: New port: Database driven web content management system (CMS)
Drupal is a free web Content Management System (CMS) that allows an
individual or a community of users to easily publish, manage and organize a
wide variety of content on a website.

Drupal is ready to go from the moment you download it. It even has an
easy-to-use web installer! The built-in functionality, combined with dozens
of freely available add-on modules, will enable features such as: Content
Management Systems, Blogs, Collaborative authoring environments, Forums,
Peer-to-peer networking, Newsletters, Podcasting, Picture galleries, File
uploads/downloads and much more.
PR:		268463