Mk/bsd.sanity.mk: Reduce parsing during package builds
This disables bsd.sanity.mk when PACKAGE_BUILDING is set such that it
only is loaded during 'make [all]' or 'make check-sanity'.  It is assumed the
package tool will manually call the target if it wants the check.  Otherwise
we get redundant logic and filesystem lookups from this file for every other
build target.

Prodded by:	mjg