net-p2p/tootle: Add new port
Tootle is a client for Mastodon, which is the world's largest
decentralized microblogging platform. The client provides a clean,
native interface, allowing the Mastodon social experience to be
seamlessly integrated into our desktop environment.

This client for Mastodon that It is characterized by three fundamental
aspects; be simple, free and open source. It should be added that this
tool is based on GTK with dark mode support.