Some Haskell ports: Mark BROKEN on aarch64.
These ports are failing on build cluster for some time. I tried reproducing
it locally and it turned out to be a strange heisenbug - the compiler crashes
when compiling different unrelated files and restarting the build makes the
problem go away.

I wasn't able to extract a minimal reproducing testcase, so didn't contact
upstream.

Mark BROKEN for now and hope for a new compiler version to work.
Reported by:	pkg-fallout