*/horde-*: Mark BROKEN
As part of removal of php74 the initial thought was to remove the horde
framework entirely. However considering the fact that the forks of horde
with php8 compatibility from https://github.com/maintaina-com will soon
be merged upstream we are not removing the ports but keeping them in
BROKEN state as the ports build perfectly while fails to run under
php80. Also mark the ports DEPRECATED in case the patches do not get
merged upstream. Also set the EXPIRATION_DATE to 2023-03-31. In case
someone is willing to convert the ports to use maintaina-com fork feel
free to update accordingly.

Approved by:	portmgr (blanket infrastructure)
Sponsored by:	Bounce Experts