net-mgmt/netdata: Enable CLOUD option by default
Even with the CLOUD option enabled, the Netdata agent does not connect
to the Netdata Cloud by default. This has to be manually enabled after
the agent is first run using the netdata-claim.sh script.

Add to pkg-message instructions an optional step for cloud integration,
with a link to the netdata-claim.sh online instructions.

Approved by:	nc (maintainer timeout)
Differential Revision:	https://reviews.freebsd.org/D37453