devel/nasm: Update to 2.16.01 (second attempt)
- The first attempt to update devel/nasm to version 2.16.01 failed due
to a bug in tmpfs with updating timestamps using ": >". In particular,
the bug is reproduced in poudriere using default configuration due to
USE_TMPFS=all.

The bug in tmpfs is being fixed, starting with commit
860399eb86cc431412bfbce0ab76c6652e5b6c07 in the src repository. Until
the commit is merged in all officially supported FreeBSD versions, the
patch files/patch-Makefile.in is necessary: it is a workaround that
replaces ": >" with "touch" in Makefile.in [1].

- The file files/patch-d85890aa.c is being removed as it is now obsolete:
the commit from which the patch was taken is already included in the
update.

- The RDOFF option has been removed as, according to the 2.16 changelog,
"Support for the rdf format has been discontinued and all the RDOFF
utilities has been removed." [2]

Because of the RDOFF option removal, the list of installed files is now
much shorter and pkg-plist has been replaced by PLIST_FILES as suggest
by portlint.

Changes: https://nasm.us/doc/nasmdocc.html

PR:		268528 [1,2]
Reported by:	yasu, Johan Ström <johan@stromnet.se> and many others [1,2]
Analyzed by:	yasu and Mark Millard <marklmi26-fbsd@yahoo.com> [1]
Tested by:	pi